Move IMDB to text_classification.py
